Move & rename Collector to Recorder
